requirements.txt
pip == 21.3.1
smbus == 1.1.post2
numpy == 1.21.4
adafruit-circuitpython-servokit == 1.3.5
rpi_ws281x == 4.3.1
mpu6050-raspberrypi == 1.1
flask == 2.0.2
turbo-flask == 0.7.0
